Fix chapters getting deleted when marking as unread from library (fixes #5755)

+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryPresenter.kt
@@ -487,7 +487,7 @@ class LibraryPresenter(
                 }
                 db.updateChaptersProgress(chapters).executeAsBlocking()
 
-                if (preferences.removeAfterMarkedAsRead()) {
+                if (read && preferences.removeAfterMarkedAsRead()) {
                     deleteChapters(manga, chapters)
                 }
             }
